additional features and fixes
Fix an issue where a conditional returned none within an alternative
parser.

Return part of a binary using an offset and length.

Floating point little/big endian to 128 bits.

Tidy up redundant default unsigned or big from scran_number:int/3
clauses.

Truncate floating point to a number of decimal digits with
scran_number:precision/1.

Zip the result of one parser with the result of another in
scran_sequence:zip/2.

Use the result of the first parser to initialise the second with
scran_sequence:followed_with/2.

Combine the result of the first parser with a second parser (also
initialised with the result of the first), typically where both are
returning maps using scran_sequence:combined_with/2.